requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
boto3
